# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API của QuickWork",
        "title": "QuickWork API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Đăng nhập hệ thống",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập",
                "parameters": [
                    {
                        "description": "Thông tin đăng nhập",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/register-enterprise": {
            "post": {
                "description": "Tiếp nhận thông tin doanh nghiệp.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký tài khoản Doanh nghiệp",
                "parameters": [
                    {
                        "description": "Payload Đăng ký Doanh nghiệp",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterEnterpriseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Thành công",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Lỗi định dạng dữ liệu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Trùng Email hoặc TaxCode",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/register-student": {
            "post": {
                "description": "Tiếp nhận thông tin cơ bản của học viên để mở tài khoản tự động và khởi tạo Profile.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký tài khoản Sinh viên",
                "parameters": [
                    {
                        "description": "Payload Đăng ký Sinh viên",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterStudentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Thành công",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Lỗi định dạng dữ liệu",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Trùng lặp Email",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Lỗi hệ thống",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "request.RegisterEnterpriseRequest": {
            "type": "object",
            "required": [
                "company_name",
                "email",
                "password",
                "tax_code"
            ],
            "properties": {
                "company_name": {
                    "type": "string",
                    "minLength": 3
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "tax_code": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10
                }
            }
        },
        "request.RegisterStudentRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string",
                    "maxLength": 11,
                    "minLength": 10
                }
            }
        }
    }
}